Rename types library, folder, and namespace to payload_types
